Merge popen-compatibility-733-6
Author: glyph, itamar, exarkun, (perhaps others sorry if I forgot you)
Reviewer: glyph, dreid, jknight
Fixes: #733

Change the reactors to, on POSIX, use `signal.set_wakeup_fd`
and `signal.siginterrupt` to support child processes without
creating the possibility for causing syscalls to fail with
EINTR.  On Python 2.5 and older, where these APIs are not
available, a new extension module is added to implement the
functionality.  If neither of these is available, revert to
the old behavior.  Also Glib2/Gtk2 reactor retain the old
behavior due to implementation difficulties.


git-svn-id: svn://svn.twistedmatrix.com/svn/Twisted/trunk@28447 bbbe8e31-12d6-0310-92fd-ac37d47ddeeb