Merge correct-file-mode-4207
Author: jkakar
Reviewers: ivank, jesstess
Fixes: #4207

Due to a bug in Python 2.7a on Windows, including multiple 'b'
characters in the mode passed to the built-in open() will cause an
error.  FilePath.open has been modified to ensure that only a single
'b' character is included in the mode passed to the built-in open()
to avoid this issue.  Callsites that pass a redundant 'b' mode have
been updated to FilePath.open no longer do so, since it ensures that
'b' is always present in the mode.  This branch also includes some
drive-by formatting cleanups and copyright updates.

See http://bugs.python.org/issue7686 for details about the bug in
Python 2.7a that necessitated this change.




git-svn-id: svn://svn.twistedmatrix.com/svn/Twisted/trunk@28793 bbbe8e31-12d6-0310-92fd-ac37d47ddeeb