Revert r29015
The new setContent uses temporarySibling, which calls requireCreate, which makes open() call create(), and create() doesn't call a mode argument, but in order to be correct it needs to take O_BINARY on Windows.  This is really a separate bug, but this fix should be merged after it is in.

Reopens: #3694



git-svn-id: svn://svn.twistedmatrix.com/svn/Twisted/trunk@29018 bbbe8e31-12d6-0310-92fd-ac37d47ddeeb