Merge less-confusing-accept-emfile-test-4221
Author: exarkun
Reviewer: jesstess, glyph
Fixes: #4221

Change one of the platform integration tests so that when it fails,
it fails with a more informative error.

Previously `test_acceptOutOfFiles` would fail with an `EBADF` if it
ran out of file descriptors before it managed to allocate any clients.
It did this because it would accidentally close the port it wanted to
accept on.

The test is changed to allocate sockets in a different order now, so
if the same condition is hit, it will fail with a boring `EMFILE` instead. 


git-svn-id: svn://svn.twistedmatrix.com/svn/Twisted/trunk@29039 bbbe8e31-12d6-0310-92fd-ac37d47ddeeb