Merge closingpipes-4626
Author: exarkun
Reviewer: jml
Fixes: #4626

Fix a race condition in `test_stdout` and `test_stderr` which would sometimes
cause the child process to exit before the parent process managed to close its
stdin.  This prevent the test from actually exercising the code path it was
trying to exercise.  The new version adds a lot of delays to the child which
should make the race go the right way almost all of the time.

A fix which completely removes the race would be nice, but appears to be much
more complicated.


git-svn-id: svn://svn.twistedmatrix.com/svn/Twisted/trunk@29940 bbbe8e31-12d6-0310-92fd-ac37d47ddeeb