Merge protocol-ssl-4854-6
Author: exarkun
Reviewer: itamar, glyph, thijs
Fixes: #4854
Refs: #4974
Refs: #5014
Refs: #4455

Add an implementation of `IReactorSSL` and `ITLSTransport` which uses the memory
BIO APIs present in pyOpenSSL 0.10 and newer.  This implementation will be preferred
by all reactors if the pyOpenSSL dependency is satisfied, otherwise the old
implementation will still be used.

This appears to have slightly better performance than the old implementation and
should avoid bugs like #4455.


git-svn-id: svn://svn.twistedmatrix.com/svn/Twisted/trunk@31537 bbbe8e31-12d6-0310-92fd-ac37d47ddeeb