Merge pubkey-check-5349
Author: antoine
Reviewer: exarkun
Fixes: #5349

Explicitly use the module name of the class of the key object
to determine what type of key it is, instead of relying on the
string representation of the class object.

This avoids a problem on Python 3 where the PyCrypto key class
becomes new-style and has a different string representation.


git-svn-id: svn://svn.twistedmatrix.com/svn/Twisted/trunk@33046 bbbe8e31-12d6-0310-92fd-ac37d47ddeeb