Merge win32er-close-5233-2
Author: exarkun
Reviewer: itamar, zseil, glyph
Fixes: #5233

win32eventreactor is now more careful in its handling of socket close
notifications.  Where previously it could miss a close notification
entirely if it arrived at the same time as the last chunk of bytes on
a connection, it will now handle that case and properly disconnect
the transport and protocol.  However, this functionality requires

pywin32 215 or later; it will be disabled with a warning if an older
version of pywin32 is installed.


git-svn-id: svn://svn.twistedmatrix.com/svn/Twisted/trunk@33049 bbbe8e31-12d6-0310-92fd-ac37d47ddeeb