Merge do-not-shutdown-sockets-5621-3
Author: exarkun
Reviewer: therve, rwall
Fixes: #5621

Remove the `shutdown(2)` call on file descriptors added to the reactor using
`IReactorSocket.adoptStreamPort`.  The socket is still closed, which will
have the same effect if and only if no other references to the socket exist
(for example, in a systemd process, where it was originally created).


git-svn-id: svn://svn.twistedmatrix.com/svn/Twisted/trunk@34225 bbbe8e31-12d6-0310-92fd-ac37d47ddeeb