Merge check-zi-version-5935-2
Author: exarkun
Reviewer: glyph
Fixes: #5935

Add a check for the required version of zope.interface to `twisted/__init__.py` and present
a meaningful exception when the requirement is not satisfied.  This improves on the previous
behavior where a strange and often confusing exception would be raised when a too-old version
of zope.interface is installed.


git-svn-id: svn://svn.twistedmatrix.com/svn/Twisted/trunk@35648 bbbe8e31-12d6-0310-92fd-ac37d47ddeeb